+++ CHANGELOG.md
@@ -19,6 +19,7 @@
 * Fix blowing up a specific bombable wall in MQ Dodongo's Cavern opening the door controlled by the silver rupees.
 * Fix a generator failure when Water Temple is MQ and `Logic Rules` are `Advanced`.
 * Fix an error in the language system when Ganon's Boss Key is on LACS, LACS condition is vanilla, and altar misc hints are enabled.
+* When pre-completed dungeons are set to specific rewards, those rewards will no longer be placed on Rauru.
 
 ## Other changes
 * Python 3.9 through 3.12 are no longer supported.